# contracts/work-platform-bff-openapi.yaml
openapi: 3.0.3
info:
  title: Work Platform BFF API Contract (Agent-Lifeform)
  version: 1.0.0
  description: >
    工作平台（统一工作台）BFF 接口契约（W2 扩展）。
    覆盖 14 模块：总览/生命体征/决策沙盘/五感/进化/协作总线/任务/对话/专家/技能/连接器/自动化/灵感/审批。
    Mock 数据必须与该契约结构完全一致 → 切换零 UI 改动（R-C09 延续）。
    设计依据：《工作平台功能设计文档》v1.0 §6。

    **统一错误信封（2026-09-18 起全平台一致）**：所有错误响应均为
    `{code, message, details}`，与 gateway-service、Java 三服务、nlp-service 完全同构；
    BFF 路由层分流：路径不存在 → 404 `AGENT_NOT_FOUND`，路径存在但方法不支持 →
    405 `AGENT_METHOD_NOT_ALLOWED`（附 `Allow` 头），请求体非法 → 400 `AGENT_BAD_REQUEST`，
    鉴权失败 → 401 `AGENT_UNAUTHORIZED` / 403 `AGENT_FORBIDDEN`。
    区分 404 与 5xx 是前端降级判定的前提（404 = 端点未实现 = planned）。
    完整异常传播链见 `docs/异常流程归纳.md` §2。
servers:
  - url: /api/wp
    description: Work Platform BFF 聚合层
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  responses:
    ErrorEnvelope:
      description: 统一错误响应（`{code, message, details}`）—— 见 info.description
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
  schemas:
    ApiResponse:
      type: object
      required: [code, data, msg]
      properties:
        code: { type: integer, example: 0 }
        data: { type: object }
        msg: { type: string, example: ok }
    ErrorEnvelope:
      type: object
      description: |
        全平台统一错误信封。2026-09-18 起 BFF（原 `{error}`）与 nlp-service
        （原 FastAPI 默认 `{detail}`）已对齐到本结构，客户端只需一套解析逻辑。
      required: [code, message, details]
      properties:
        code:
          type: string
          description: 统一错误码（R1-08）
          enum:
            - AGENT_BAD_REQUEST
            - AGENT_UNAUTHORIZED
            - AGENT_FORBIDDEN
            - AGENT_NOT_FOUND
            - AGENT_METHOD_NOT_ALLOWED
            - AGENT_CONFLICT
            - AGENT_DUPLICATE
            - AGENT_TIMEOUT
            - AGENT_BUS_UNAVAILABLE
            - AGENT_UPSTREAM_UNAVAILABLE
            - AGENT_INTERNAL_ERROR
            # Phase 5 四肢期：工具域专属码（tool-executor `ErrorCode` 同源副本）
            # 「被拒 ≠ 服务不可用」：前四个码都是**已落审计的真实调用结果**，
            # 出现在 200 + success:false 的执行响应里，而不是 5xx。
            - AGENT_TOOL_NOT_ALLOWED
            - AGENT_TOOL_ARGS_INVALID
            - AGENT_TOOL_ARGS_BLOCKED
            - AGENT_TOOL_EXEC_FAILED
            - AGENT_TOOL_CIRCUIT_OPEN
            - AGENT_SANDBOX_REJECTED
          example: AGENT_NOT_FOUND
        message: { type: string, example: 接口不存在：/api/wp/not-exist }
        details:
          type: object
          description: 上下文明细（字段级原因 / method+supported / path 等）
          additionalProperties: true
          example: { path: /api/wp/not-exist }
...
paths:
...
  /session:
    x-wp-status: implemented
    post:
      summary: 'R4-01 创建会话（代理 session-manager，FSM NEW→ACTIVE）'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '405':
          $ref: '#/components/responses/ErrorEnvelope'
  /session/stats:
    x-wp-status: implemented
    get:
      summary: 'R-C04 会话真相：活跃会话数与意图分布（session-manager Redis SCAN 实时汇总）'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '405':
          $ref: '#/components/responses/ErrorEnvelope'
  /session/{session_id}:
    x-wp-status: implemented
    delete:
      summary: 'R4-01 关闭会话（FSM → CLOSED）'
      parameters:
        - { name: session_id, in: path, required: true, schema: { type: string } }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '405':
          $ref: '#/components/responses/ErrorEnvelope'
  /session/{session_id}/ask:
    x-wp-status: implemented
    post:
      summary: 'R4-01/02 会话问答：FSM 校验 → 意图 → 大脑层（多轮上下文取自 Redis 持久化）'
      parameters:
        - { name: session_id, in: path, required: true, schema: { type: string } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [question]
              properties:
                question: { type: string }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '400':
          $ref: '#/components/responses/ErrorEnvelope'
        '404':
          $ref: '#/components/responses/ErrorEnvelope'
        '409':
          $ref: '#/components/responses/ErrorEnvelope'
        '405':
          $ref: '#/components/responses/ErrorEnvelope'
  /session/{session_id}/context:
    x-wp-status: implemented
    get:
      summary: 'R4-02 多轮上下文回放（重启可恢复的证据来源）'
      parameters:
        - { name: session_id, in: path, required: true, schema: { type: string } }
        - { name: turns, in: query, required: false, schema: { type: integer, default: 10 } }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '405':
          $ref: '#/components/responses/ErrorEnvelope'
...